Support Ensime server 2.0. Closes #300
- Rebase on latest master
- Factor out response handlers to separate mixin/file
- Create EnsimeClient for V1 or V2 based on server_v2 flag
- Setup EnsimeLauncher according to server_v2 flag also
