+++ ensime_shared/launcher.py
@@ -55,11 +55,13 @@ class EnsimeProcess(object):
 
 
 class EnsimeLauncher(object):
-    ENSIME_VERSION = '1.0.0'
+    ENSIME_V1 = '1.0.0'
+    ENSIME_V2 = '2.0.0-SNAPSHOT'
     SBT_VERSION = '0.13.11'
 
-    def __init__(self, vim, config_path, base_dir=BOOTSTRAPS_ROOT):
+    def __init__(self, vim, config_path, server_v2, base_dir=BOOTSTRAPS_ROOT):
         self.vim = vim
+        self.ensime_version = self.ENSIME_V2 if server_v2 else self.ENSIME_V1
         self._config_path = os.path.abspath(config_path)
         self.config = self.parse_config(self._config_path)
         self.base_dir = os.path.abspath(base_dir)
@@ -219,7 +221,7 @@ saveClasspathTask := {
 }"""
         replace = {
             "scala_version": self.config['scala-version'],
-            "version": self.ENSIME_VERSION,
+            "version": self.ensime_version,
             "classpath_file": self.classpath_file,
         }
